Refinement

+++ es-core/src/watchers/SyncthingWatcher.h
@@ -27,8 +27,8 @@ private:
 	std::vector<std::string> mDirtyDevices;
 
 	bool mkillNotificationInNextCycle = false;
-	int mCurrentTransferNeededFiles = 0;
-	int mTotalBytesTransferred = 0;
+	int64_t mCurrentTransferNeededFiles = 0;
+    int64_t mTotalBytesTransferred = 0;
 
 	std::string toSyncedDevicesNameString(const std::vector<std::string>& deviceNames);
 };
